+++ api/src/main/kotlin/handler/TimetableHandler.kt
@@ -5,7 +5,7 @@ import com.wafflestudio.snu4t.common.exception.InvalidPathParameterException
 import com.wafflestudio.snu4t.middleware.SnuttRestApiDefaultMiddleware
 import com.wafflestudio.snu4t.timetables.dto.request.TimetableAddRequestDto
 import com.wafflestudio.snu4t.timetables.dto.request.TimetableModifyRequestDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableModifyThemeRequestDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableModifyThemeRequestDto
 import com.wafflestudio.snu4t.timetables.service.TimetableService
 import kotlinx.coroutines.flow.toList
 import org.springframework.stereotype.Component
