+++ api/src/main/kotlin/handler/TimetableThemeHandler.kt
@@ -5,12 +5,12 @@ import com.wafflestudio.snu4t.common.dto.OkResponse
 import com.wafflestudio.snu4t.common.enum.BasicThemeType
 import com.wafflestudio.snu4t.common.exception.InvalidPathParameterException
 import com.wafflestudio.snu4t.middleware.SnuttRestApiDefaultMiddleware
-import com.wafflestudio.snu4t.timetables.dto.TimetableThemeDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableThemeAddRequestDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableThemeDownloadRequestDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableThemeModifyRequestDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableThemePublishRequestDto
-import com.wafflestudio.snu4t.timetables.service.TimetableThemeService
+import com.wafflestudio.snu4t.theme.dto.TimetableThemeDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableThemeAddRequestDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableThemeDownloadRequestDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableThemeModifyRequestDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableThemePublishRequestDto
+import com.wafflestudio.snu4t.theme.service.TimetableThemeService
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.server.ServerRequest
 import org.springframework.web.reactive.function.server.awaitBody
