+++ api/src/main/kotlin/router/docs/TimetableDocs.kt
@@ -5,7 +5,7 @@ import com.wafflestudio.snu4t.timetables.dto.request.CustomTimetableLectureAddLe
 import com.wafflestudio.snu4t.timetables.dto.request.TimetableAddRequestDto
 import com.wafflestudio.snu4t.timetables.dto.request.TimetableLectureModifyLegacyRequestDto
 import com.wafflestudio.snu4t.timetables.dto.request.TimetableModifyRequestDto
-import com.wafflestudio.snu4t.timetables.dto.request.TimetableModifyThemeRequestDto
+import com.wafflestudio.snu4t.theme.dto.request.TimetableModifyThemeRequestDto
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.enums.ParameterIn
