+++ core/src/main/kotlin/theme/data/ColorSet.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.snu4t.timetables.data
+package com.wafflestudio.snu4t.theme.data
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import org.springframework.data.mongodb.core.mapping.Field
