+++ core/src/main/kotlin/theme/dto/request/TimetableThemeAddRequestDto.kt
@@ -1,0 +1,8 @@
+package com.wafflestudio.snu4t.theme.dto.request
+
+import com.wafflestudio.snu4t.theme.data.ColorSet
+
+data class TimetableThemeAddRequestDto(
+    val name: String,
+    val colors: List<ColorSet>,
+)
