+++ core/src/main/kotlin/theme/dto/request/TimetableThemeDownloadRequestDto.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.snu4t.timetables.dto.request
+package com.wafflestudio.snu4t.theme.dto.request
 
 class TimetableThemeDownloadRequestDto (
     val name: String,
