+++ core/src/main/kotlin/theme/dto/request/TimetableThemeModifyRequestDto.kt
@@ -1,0 +1,8 @@
+package com.wafflestudio.snu4t.theme.dto.request
+
+import com.wafflestudio.snu4t.theme.data.ColorSet
+
+data class TimetableThemeModifyRequestDto(
+    val name: String?,
+    val colors: List<ColorSet>?,
+)
