+++ core/src/main/kotlin/timetables/dto/request/CustomTimetableLectureAddLegacyRequestDto.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.snu4t.timetables.dto.request
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.wafflestudio.snu4t.timetables.data.ColorSet
+import com.wafflestudio.snu4t.theme.data.ColorSet
 import com.wafflestudio.snu4t.timetables.data.TimetableLecture
 
 data class CustomTimetableLectureAddLegacyRequestDto(
