+++ core/src/main/kotlin/timetables/dto/request/TimetableLectureModifyLegacyRequestDto.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.snu4t.timetables.dto.request
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.wafflestudio.snu4t.timetables.data.ColorSet
+import com.wafflestudio.snu4t.theme.data.ColorSet
 
 data class TimetableLectureModifyLegacyRequestDto(
     @JsonProperty("course_title")
