--- core/src/main/kotlin/timetables/dto/request/TimetableThemeModifyRequestDto.kt
@@ -1,8 +1,0 @@
-package com.wafflestudio.snu4t.timetables.dto.request
-
-import com.wafflestudio.snu4t.timetables.data.ColorSet
-
-data class TimetableThemeModifyRequestDto(
-    val name: String?,
-    val colors: List<ColorSet>?,
-)
